pyproject.toml
[project]
name = "tanamanhias"
version = "0.0.1"
requires-python = ">=3.11,<3.12"
dependencies = [
  "streamlit==1.37.1",
  "tensorflow-cpu==2.16.1",
  "numpy==1.26.4",
  "pillow==10.4.0",
  "protobuf==4.25.3",
]
